Have managed to work out some of the connection issues, but still not there.
Even with all the proper annotation AND setting the full application.properties file back up it will still not connect all the repositories.

+++ tipsy.security/src/main/java/net/treXis/tipsy/security/model/Authorities.java
@@ -2,16 +2,16 @@ package net.treXis.tipsy.security.model;
 
 public class Authorities {
 
-    private int user_account;
+    private Long user_account;
     private String username;
     private String authority;
 
 
-    public int getUser_account() {
+    public Long getUser_account() {
         return user_account;
     }
 
-    public void setUser_account(int user_account) {
+    public void setUser_account(Long user_account) {
         this.user_account = user_account;
     }
